Revert "refactor HandlerFactory"
This reverts commit 04cfbc77a8eeacd6e3b5cd39d614d7195a074350.
